+++ src/app/addProject/page.tsx
@@ -4,7 +4,7 @@ import { getServerSession } from 'next-auth';
 import { prisma } from '@/lib/prisma';
 import AddProjectForm from '@/components/AddProjectForm';
 import { loggedInProtectedPage } from '@/lib/page-protection';
-import { authOptions } from '../api/auth/[...nextauth]/route';
+import authOptions from '@/app/auth/authOptions/authOptions';
 
 const AddProjectPage = async () => {
   const session = await getServerSession(authOptions);
