+++ src/app/api/route.ts
@@ -1,6 +1,6 @@
 import { getServerSession } from 'next-auth/next';
 import { NextResponse } from 'next/server';
-import { authOptions } from './auth/[...nextauth]/route';
+import authOptions from '@/app/auth/authOptions/authOptions';
 
 export default async function GET() {
   const session = await getServerSession(authOptions);
