+++ src/app/auth/authOptions/authOptions.ts
@@ -1,0 +1,79 @@
+import { compare } from 'bcrypt';
+import { type NextAuthOptions } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import { prisma } from '@/lib/prisma';
+
+const authOptions:NextAuthOptions = {
+  session: {
+    strategy: 'jwt',
+  },
+  providers: [
+    CredentialsProvider({
+      name: 'Email and Password',
+      credentials: {
+        email: {
+          label: 'Email',
+          type: 'email',
+          placeholder: 'john@foo.com',
+        },
+        password: { label: 'Password', type: 'password' },
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials.password) {
+          return null;
+        }
+        const user = await prisma.user.findUnique({
+          where: {
+            email: credentials.email,
+          },
+        });
+        if (!user) {
+          return null;
+        }
+
+        const isPasswordValid = await compare(credentials.password, user.password);
+        if (!isPasswordValid) {
+          return null;
+        }
+
+        return {
+          id: `${user.id}`,
+          email: user.email,
+          randomKey: user.role,
+        };
+      },
+    }),
+  ],
+  pages: {
+    signIn: '/auth/signin',
+    signOut: '/auth/signout',
+    //   error: '/auth/error',
+    //   verifyRequest: '/auth/verify-request',
+    //   newUser: '/auth/new-user'
+  },
+  callbacks: {
+    session: ({ session, token }) => ({
+      // console.log('Session Callback', { session, token })
+      ...session,
+      user: {
+        ...session.user,
+        id: token.id,
+        randomKey: token.randomKey,
+      },
+    }),
+    jwt: ({ token, user }) => {
+      // console.log('JWT Callback', { token, user })
+      if (user) {
+        const u = user as unknown as any;
+        return {
+          ...token,
+          id: u.id,
+          randomKey: u.randomKey,
+        };
+      }
+      return token;
+    },
+  },
+};
+
+export default authOptions;
